csskit: Modularise commands
This should help as the command complexity grows
